[PCC-45] Refactor: Change all Floats to Doubles

+++ app/src/main/java/com/github/komidawi/pizzacostcalculator/screen/list/BindingAdapters.kt
@@ -26,5 +26,5 @@ fun TextView.setPizzaPriceDisplay(pizza: PizzaEntity?) {
 
 @BindingAdapter("listPizzaRatioDisplay")
 fun TextView.setListPizzaRatioDisplay(pizza: PizzaEntity?) {
-    pizza?.let { text = String.format("%.0f", pizza.ratio.toFloat()) }
+    pizza?.let { text = String.format("%.0f", pizza.ratio.toDouble()) }
 }
